requirements.txt
python-dotenv
aiogram==3.0.0
gunicorn==21.2.0
